refs #14 Add tag pin button in detail
