+++ src/components/detail/Detail.tsx
@@ -1,12 +1,9 @@
 import { useRouter } from 'next/router'
 import { Dispatch, useEffect, useState } from 'react'
-import { BsX, BsChevronLeft } from 'react-icons/bs'
-import { Animation, Container, Header, FlexboxGrid, Button } from 'rsuite'
-import { Icon } from '@rsuite/icons'
+import { Animation, Container } from 'rsuite'
 
 import Status from './Status'
 import Profile from './Profile'
-import { useTranslation } from 'react-i18next'
 import { MegalodonInterface } from 'megalodon'
 import TagDetail from './Tag'
 
@@ -18,8 +15,6 @@ type Props = {
 }
 
 const Detail: React.FC<Props> = props => {
-  const { t } = useTranslation()
-
   const [target, setTarget] = useState<'status' | 'profile' | 'tag' | null>(null)
   const router = useRouter()
 
@@ -35,14 +30,6 @@ const Detail: React.FC<Props> = props => {
     }
   }, [router.query])
 
-  const back = () => {
-    router.back()
-  }
-
-  const close = () => {
-    router.push({ query: {} })
-  }
-
   return (
     <Animation.Transition
       in={target !== null}
@@ -54,21 +41,6 @@ const Detail: React.FC<Props> = props => {
       {(p, ref) => (
         <div {...p} ref={ref}>
           <Container className="profile" style={{ height: '100%', borderLeft: '1px solid var(--rs-gray-600)', overflow: 'hidden' }}>
-            <Header style={{ backgroundColor: 'var(--rs-gray-700)' }}>
-              <FlexboxGrid justify="space-between">
-                <FlexboxGrid.Item>
-                  <Button appearance="link" onClick={back}>
-                    <Icon as={BsChevronLeft} style={{ fontSize: '1.4em' }} />
-                    {t('detail.back')}
-                  </Button>
-                </FlexboxGrid.Item>
-                <FlexboxGrid.Item>
-                  <Button appearance="link" onClick={close}>
-                    <Icon as={BsX} style={{ fontSize: '1.4em' }} />
-                  </Button>
-                </FlexboxGrid.Item>
-              </FlexboxGrid>
-            </Header>
             {target === 'status' && (
               <Status
                 openMedia={(media: Array<Entity.Attachment>, index: number) =>
